autoApproveSync: also run when PR labeled.
As explained in https://github.com/nextcloud/android-library/pull/1060#issuecomment-1423822136, the sync
action adds the label *after* opening the PR, so when this workflow runs the label isn't there yet,
and the auto-approve action will be skipped.

This fixes that by also running this workflow when the PR is labeled.

Also added a concurrency group to avoid running this in parallel.

Signed-off-by: Álvaro Brey <alvaro.brey@nextcloud.com>